@@ -59,7 +59,7 @@ public class ParallelTestClassRunner : XunitTestClassRunner
     {
         while (true)
         {
-            if (ex is not TargetInvocationException tiex || tiex.InnerException == null)
+            if (ex is not TargetInvocationException tiex || tiex.InnerException is null)
                 return ex;
 
             ex = tiex.InnerException;
