@@ -2,19 +2,12 @@ using Xunit;
 
 namespace Meziantou.Xunit.ParallelTestFramework.Tests;
 
-public class BlockingTheoryTests : IClassFixture<ConcurrencyFixture>
+public class BlockingTheoryTests(ConcurrencyFixture fixture) : IClassFixture<ConcurrencyFixture>
 {
-    private readonly ConcurrencyFixture fixture;
-
-    public BlockingTheoryTests(ConcurrencyFixture fixture)
-    {
-        this.fixture = fixture;
-    }
-
     [Theory]
     [InlineData(1)]
     [InlineData(2)]
-    public void Theory(int value)
+    public void Theory(int _)
     {
         Assert.Equal(2, fixture.CheckConcurrencyAsync().Result);
     }
