@@ -1,21 +1,14 @@
-﻿using Xunit;
+using Xunit;
 
 namespace Meziantou.Xunit.ParallelTestFramework.Tests;
 
 [Collection("CollectionAttribute_SequentialTheoryTests")]
-public class CollectionAttributeSequentialTheoryTests : IClassFixture<ConcurrencyFixture>
+public class CollectionAttributeSequentialTheoryTests(ConcurrencyFixture fixture) : IClassFixture<ConcurrencyFixture>
 {
-    private readonly ConcurrencyFixture fixture;
-
-    public CollectionAttributeSequentialTheoryTests(ConcurrencyFixture fixture)
-    {
-        this.fixture = fixture;
-    }
-
     [Theory]
     [InlineData(1)]
     [InlineData(2)]
-    public async Task Theory(int value)
+    public async Task Theory(int _)
     {
         Assert.Equal(1, await fixture.CheckConcurrencyAsync().ConfigureAwait(false));
     }
