@@ -2,17 +2,17 @@ namespace Meziantou.Xunit.ParallelTestFramework.Tests;
 
 public class CollectionConcurrencyFixture
 {
-    private int concurrency;
+    private int _concurrency;
 
     public async Task<int> CheckConcurrencyAsync()
     {
-        Interlocked.Increment(ref concurrency);
+        Interlocked.Increment(ref _concurrency);
         await Task.Delay(TimeSpan.FromMilliseconds(1000)).ConfigureAwait(false);
 
-        var overlap = concurrency;
+        var overlap = _concurrency;
 
         await Task.Delay(TimeSpan.FromMilliseconds(1000)).ConfigureAwait(false);
-        Interlocked.Decrement(ref concurrency);
+        Interlocked.Decrement(ref _concurrency);
 
         return overlap;
     }
