@@ -2,20 +2,13 @@ using Xunit;
 
 namespace Meziantou.Xunit.ParallelTestFramework.Tests;
 
-public class MemberDataAttributeTests : IClassFixture<ConcurrencyFixture>
+public class MemberDataAttributeTests(ConcurrencyFixture fixture) : IClassFixture<ConcurrencyFixture>
 {
-    private readonly ConcurrencyFixture fixture;
-
     public static TheoryData<int> GetData() => new() { { 1 }, { 2 } };
-
-    public MemberDataAttributeTests(ConcurrencyFixture fixture)
-    {
-        this.fixture = fixture;
-    }
 
     [Theory]
     [MemberData(nameof(GetData), DisableDiscoveryEnumeration = true)]
-    public async Task Theory(int value)
+    public async Task Theory(int _)
     {
         Assert.Equal(1, await fixture.CheckConcurrencyAsync().ConfigureAwait(false));
     }
