@@ -2,19 +2,12 @@ using Xunit;
 
 namespace Meziantou.Xunit.ParallelTestFramework.Tests;
 
-public class ParallelTheoryTests : IClassFixture<ConcurrencyFixture>
+public class ParallelTheoryTests(ConcurrencyFixture fixture) : IClassFixture<ConcurrencyFixture>
 {
-    private readonly ConcurrencyFixture fixture;
-
-    public ParallelTheoryTests(ConcurrencyFixture fixture)
-    {
-        this.fixture = fixture;
-    }
-
     [Theory]
     [InlineData(1)]
     [InlineData(2)]
-    public async Task Theory(int value)
+    public async Task Theory(int _)
     {
         Assert.Equal(2, await fixture.CheckConcurrencyAsync().ConfigureAwait(false));
     }
