@@ -3,15 +3,8 @@ using Xunit;
 namespace Meziantou.Xunit.ParallelTestFramework.Tests;
 
 [Collection("Sequential")]
-public class SequentialCollectionAttributeTests : IClassFixture<ConcurrencyFixture>
+public class SequentialCollectionAttributeTests(ConcurrencyFixture fixture) : IClassFixture<ConcurrencyFixture>
 {
-    private readonly ConcurrencyFixture fixture;
-
-    public SequentialCollectionAttributeTests(ConcurrencyFixture fixture)
-    {
-        this.fixture = fixture;
-    }
-
     [Fact]
     public async Task Fact1()
     {
