@@ -2,20 +2,13 @@ using Xunit;
 
 namespace Meziantou.Xunit.ParallelTestFramework.Tests;
 
-public class SequentialTheoryTests : IClassFixture<ConcurrencyFixture>
+public class SequentialTheoryTests(ConcurrencyFixture fixture) : IClassFixture<ConcurrencyFixture>
 {
-    private readonly ConcurrencyFixture fixture;
-
-    public SequentialTheoryTests(ConcurrencyFixture fixture)
-    {
-        this.fixture = fixture;
-    }
-
     [Theory]
     [InlineData(1)]
     [InlineData(2)]
     [DisableParallelization]
-    public async Task Theory(int value)
+    public async Task Theory(int _)
     {
         Assert.Equal(1, await fixture.CheckConcurrencyAsync().ConfigureAwait(false));
     }
